APIs in Spring

+++ src/main/java/com/demo/first/app/UserController.java
@@ -28,6 +28,8 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public String deleteUser(@PathVariable int id){
+        if (userDb.containsKey(id))
+            return "User not found!";
         userDb.remove(id);
         return "User Deleted!";
     }
